Adding webpack configutration
