+++ index.html
@@ -12,7 +12,7 @@
     <link rel="stylesheet" href="https://use.fontawesome.com/releases/v5.5.0/css/all.css" integrity="sha384-B4dIYHKNBt8Bc12p+WXckhzcICo0wtJAoU8YZTY5qE0Id1GSseTk6S+L3BlXeVIU" crossorigin="anonymous">
 
     <!-- CUSTOM STYLES -->
-    <link rel="stylesheet" href="style.css">
+    <link rel="stylesheet" href="./dist/style.css">
 </head>
 
 <body>
@@ -180,7 +180,7 @@
     </section>
 
 
-    <script src="custom.js"></script>
+    <script src="./dist/bundle.js"></script>
 </body>
 
 </html>
