+++ package.json
@@ -1,0 +1,32 @@
+{
+  "name": "Budget-App",
+  "version": "1.0.0",
+  "description": "This application was created based on Jonas Schmedtmann course on Udemy website (course title: \"The Complete JavaScript Course 2018: Build Real Projects!\").",
+  "main": "custom.js",
+  "scripts": {
+    "build": "webpack"
+  },
+  "repository": {
+    "type": "git",
+    "url": "git+https://github.com/krzysztofpodmokly/Budget-App.git"
+  },
+  "keywords": [],
+  "author": "",
+  "license": "ISC",
+  "bugs": {
+    "url": "https://github.com/krzysztofpodmokly/Budget-App/issues"
+  },
+  "homepage": "https://github.com/krzysztofpodmokly/Budget-App#readme",
+  "devDependencies": {
+    "babel-core": "^6.26.3",
+    "babel-loader": "^6.4.1",
+    "babel-preset-env": "^1.7.0",
+    "css-loader": "^2.1.1",
+    "extract-text-webpack-plugin": "^2.0.0-beta.4",
+    "file-loader": "^3.0.1",
+    "image-webpack-loader": "^4.6.0",
+    "style-loader": "^0.23.1",
+    "url-loader": "^1.1.2",
+    "webpack": "^2.2.0-rc.0"
+  }
+}
